Infilect_Assesment/Infilect_Weather_API_Assignment/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
